requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
starlette==0.27.0
deepgram-sdk==3.2.6
groq==0.4.1
pydantic-settings==2.0.3
openai==1.3.6
python-dotenv==1.0.0
websockets==12.0
httpx==0.25.2
cryptography
anthropic==0.7.8
pyjwt==2.8.0
bcrypt==4.1.2
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
pymongo==4.6.0
google-generativeai==0.8.3
stripe
certifi